+++ tests/unit/unit_t_db_base.py
@@ -169,7 +169,6 @@ class UnitTestDbBase(unittest.TestCase):
                 timeout=timeout
             )
 
-
     def tearDown(self):
         """
         Ensure the client is new for each test